frrmanager: rewrite to use frr-reload.py
This ensures that the agent will only send configuration commands to FRR
when there are actual changes to be made in FRR's running configuration,
and also fixes cases where certain configuration lines might be
inappropriately left behind in the FRR running configuration (e.g.,
after deconfiguring underlay leaking).

Since FRR is such a massive pain to configure in an automated manner,
import the frr-reload.py script as a library. Most of the dirty work of
figuring out what steps are required to go from the running
configuration to the target configuration has already been done there,
so no need to re-invent the wheel.