feat: conditional advertisement of subnet prefixes
Implement a new feature that allows the operator to control whether or
not the directly connected subnet prefixes on a network are a being
advertised in BGP or not.

If all the IP addresses in use on a network are known in advance (i.e.,
associated with ports in the OpenStack database), then advertising the
subnet prefixes is pointless, and will only attract junk traffic
destined for unused IP addresses. However it is necessary if there are
*any* IP addresses active on the network not associated with ports
(note that being present in `allowed_address_pairs` do not count here).

Also refactor out the code that pushes config to FRR in a separate
define, to prevent repeating it thrice.