Make it possible to disable distributed FIPs
It rurns out that a port on a tenant network that has a floating IP
assigned AND is also a member of an Octavia load balancer end out in a
state where the database does not match the reality.

Specifically, even though the database would have you know that the
floating IP is distributed to compute node X using MAC address Y, in
reality it is centralised on the node where the CR-LRP port is located,
using the MAC address of the CR-LRP port. The code that makes this
happens is here:

https://opendev.org/openstack/neutron/src/branch/stable/2024.1/neutron/plugins/ml2/drivers/ovn/mech_driver/ovsdb/ovn_client.py#L1061-L1079

In this situation, the EVPN Agent is unaware of the fact that the
floating IP has been centralised will program the neighbour entry for
floating IP (and static host route when l3vni=0) as if it was
distributed, which results in a blackhole.

I haven't found out  how to detect and exclude these ports only from
being pre-programmed, so for now the workaround is an all-or-nothing
approach, allowing for disabling pre-programming neigh entries (and
static host routes) for floating IPs completely. In this mode, ARP
requests are necessary in order to discover the location and MAC address
of the floating IPs, which means this mode is fundamentally incompatible
with disabling advertise_connected for networks with floating IPs in the
database.